+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 * Made the cooldown for doob/dog longer.
     - Someone suggested in the [Doob Support Discord](https://discord.gg/hgQTTU7) to make a cooldown on the command, little did they know, it had a command, appearently it wasn't long enough so I have increased it.
         - You can only use it 2 times every 5 seconds.
+* Added owner only doob/nitrogiveaway command to tell users on how Doob's Nitro Classic giveaways work.
 
 ## v2.0.7
 * Added doob/support - Gives link to the Doob support server.